requirements.txt
djangorestframework
pyyaml
requests
django-cors-headers
django==4.2.0